+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add props on `ProjectCard` and `CrowdfundingCard` components.
+
 ## [23.2.1] - 2018-09-04
 
 Fix: